+++ src/main/java/com/example/demo1/service/TicketService.java
@@ -64,6 +64,9 @@ public class TicketService {
     }
 
     public void deleteTicket(Long id) {
+        if (!ticketRepo.existsById(id)) {
+            throw new RuntimeException(NOTFOUND);
+        }
         ticketRepo.deleteById(id);
     }
 
@@ -78,6 +81,5 @@ public class TicketService {
         }
 
     }
-
 
 }
